chore: rename old packages
