+++ packages/eslint-config-legacy-typescript/CHANGELOG.md
