+++ packages/eslint-config-legacy-typescript/README.md
