+++ packages/eslint-config-legacy-typescript/index.js
