+++ packages/eslint-config-legacy-typescript/package.json
