+++ packages/eslint-config-legacy/README.md
