+++ packages/eslint-config-legacy/test/basic.test.js
