+++ packages/eslint-config-legacy/test/validate-config.test.js
