--- packages/eslint-config-next/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@nuxt/eslint-config",
-  "version": "0.1.0",
-  "description": "ESLint config for Nuxt projects",
-  "repository": "nuxt/eslint-config",
-  "license": "MIT",
-  "main": "index.js",
-  "exports": {
-    ".": "./index.js"
-  },
-  "files": [
-    "index.js"
-  ],
-  "dependencies": {
-    "@rushstack/eslint-patch": "^1.2.0",
-    "@typescript-eslint/eslint-plugin": "^5.45.0",
-    "@typescript-eslint/parser": "^5.45.0",
-    "eslint-plugin-vue": "^9.8.0",
-    "typescript": "^4.9.3"
-  },
-  "devDependencies": {
-    "eslint": "^8.29.0"
-  },
-  "peerDependencies": {
-    "eslint": "^8.29.0"
-  }
-}
